+++ test/v1/index.html
@@ -1,0 +1,43 @@
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>L.Control.Layers.Tree Tests (v1)</title>
+
+    <link rel="stylesheet" href="../../node_modules/mocha/mocha.css" />
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" />
+    <link rel="stylesheet" href="../../L.Control.Layers.Tree.css" />
+
+    <script src="../../node_modules/chai/chai.js"></script>
+    <script src="../../node_modules/happen/happen.js"></script>
+    <script src="../../node_modules/mocha/mocha.js"></script>
+
+    <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
+    <script src="../../L.Control.Layers.Tree.js"></script>
+
+    <style>
+        .map {
+            width: 400px;
+            height: 400px;
+        }
+    </style>
+</head>
+<body>
+    <div id="map" class="map"></div>
+    <div id="mocha"></div>
+
+    <script>
+        mocha.setup('bdd');
+        mocha.globals(['L']);
+
+        // mouse events used in v1
+        globalThis.MOVE_OVER_TYPE = {type: 'mouseover'};
+        globalThis.MOVE_OUT_TYPE = {type: 'mouseout'};
+    </script>
+
+    <script src="../spec.js"></script>
+
+    <script>
+        mocha.run();
+    </script>
+</body>
+</html>
